% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baixar_dados_covid.R
\name{baixar_dados_covid}
\alias{baixar_dados_covid}
\title{Baixa o CSV da base da covid em uma pasta}
\usage{
baixar_dados_covid(pasta)
}
\arguments{
\item{pasta}{pasta onde salvarei o arquivo}
}
\description{
Baixa o CSV da base da covid em uma pasta
}
